+++ gui/threedi_plugin_grid_result_dialog.py
@@ -5,7 +5,7 @@ import os
 
 from threedi_results_analysis.utils.constants import TOOLBOX_QGIS_SETTINGS_GROUP
 from qgis.PyQt import QtWidgets, uic
-from qgis.PyQt.QtCore import pyqtSignal, pyqtSlot, QModelIndex
+from qgis.PyQt.QtCore import pyqtSignal, pyqtSlot, QModelIndex, Qt
 from qgis.PyQt.QtWidgets import QAbstractItemView
 from qgis.core import QgsSettings
 from qgis.PyQt.QtGui import QStandardItemModel, QStandardItem
@@ -199,6 +199,7 @@ class ThreeDiPluginGridResultDialog(QtWidgets.QDialog, FORM_CLASS):
             self.messageLabel.setText("Please set your 3Di working directory in the 3Di Models & Simulations settings to be able to load computational grids and results from your 3Di working directory.")
             return
 
+        rows = []
         local_schematisations = list_local_schematisations(threedi_working_dir, use_config_for_revisions=False)
         for schematisation_id, local_schematisation in local_schematisations.items():
             # Iterate over revisions
@@ -209,6 +210,7 @@ class ThreeDiPluginGridResultDialog(QtWidgets.QDialog, FORM_CLASS):
                     schema_item = QStandardItem(local_schematisation.name)
                     schema_item.setEditable(False)
                     revision_item = QStandardItem(str(revision_number))
+                    revision_item.setData(revision_number, Qt.UserRole)
                     revision_item.setEditable(False)
                     # We'll store the grid folder with the revision item for fast retrieval
                     revision_item.setData(local_revision.grid_dir)
@@ -216,17 +218,23 @@ class ThreeDiPluginGridResultDialog(QtWidgets.QDialog, FORM_CLASS):
                     result_item.setEditable(False)
                     # We'll store the result folder with the result_item for fast retrieval
                     result_item.setData(os.path.join(local_revision.results_dir, result_dir))
-                    self.model.appendRow([schema_item, revision_item, result_item])
+                    rows.append([schema_item, revision_item, result_item])
 
                 # In case no results are present, but a gridadmin is present, we still add the grid, but without result item
                 if num_of_results == 0 and os.path.exists(os.path.join(local_revision.grid_dir, "gridadmin.h5")):
                     schema_item = QStandardItem(local_schematisation.name)
                     schema_item.setEditable(False)
                     revision_item = QStandardItem(str(revision_number))
+                    revision_item.setData(revision_number, Qt.UserRole)
                     revision_item.setEditable(False)
                     # We'll store the grid folder with the revision item for fast retrieval
                     revision_item.setData(local_revision.grid_dir)
-                    self.model.appendRow([schema_item, revision_item])
+                    rows.append([schema_item, revision_item])
+
+        # Sort table rows by revision number using the UserRole
+        rows.sort(key=lambda x: x[1].data(Qt.UserRole))
+        for row in rows:
+            self.model.appendRow(row)
 
         for i in range(len(self.header_labels)):
             self.tableView.resizeColumnToContents(i)